Fix auth URL
Blood Google! All of a sudden, it started throwing 500 on an older authUrl. I spent a better part of my Saturday troubleshooting this. In my attempts to debug, I had to complete recreate an Elm web app to check what exactly was causing the 500 on the original authUrl. Fucking hell.